+++ src/components/dialog/GitConfigurationDialog.vue
@@ -1,0 +1,23 @@
+<template>
+  <default-dialog dialog-key="GitProvider" data-cy="git-settings-dialog">
+    <template v-slot:title>
+      <q-icon color="primary" name="fa-brands fa-git-alt" />
+      {{ $t('page.modelizer.settings.gitProvider.title') }}
+    </template>
+    <template v-slot:default>
+      <git-form
+          :project-name="projectName"
+          @project-git:save="DialogEvent.next({ type: 'close', key: 'GitProvider' })"/>
+    </template>
+  </default-dialog>
+</template>
+
+<script setup>
+import DialogEvent from 'src/composables/events/DialogEvent';
+import DefaultDialog from 'components/dialog/DefaultDialog';
+import GitForm from 'components/form/GitForm';
+
+defineProps({
+  projectName: String,
+});
+</script>