+++ changelog.md
@@ -16,5 +16,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add Modelizer page
 - Add the component definitions panel on the modelizer page
 - Add monaco editor to modelizer text view
+- Update git configuration for a project
 
 [unreleased]: https://github.com/ditrit/leto-modelizer/blob/main/changelog.md#unreleased